+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # :arrows_clockwise: Microsoft AD As Built Report Changelog
 
-## [0.7.4] - 2022-xx-xx
+## [0.7.5] - 2022-08-06
+
+### Added
+
+- Added SYSVOL/NETLOGON folder content status
+  - Added Health Check for malicious/unessential file extensions
+- Added Domain Controller SRV Records Status
+  - Added Health Check for SRV Records Status
+- Added Health Check for Unsupported Operating System findings in the Active Directory Domain
+### Changed
+
+- Allowed the Forest Root Domain to be the fisrt Domain in the report
+- Improved Sites Replication (repadmin) section
+
+## [0.7.4] - 2022-07-29
+
 ### Changed
 
 - Access well known groups via SID to include international names and expand them to localized group names.
@@ -9,6 +24,7 @@
 ### Fixed
 
 - Fixes [#42](https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD/issues/42)
+
 ## [0.7.3] - 2022-05-13
 
 ### Added